[WIP] Adding basic virtual FS functionality
Note that as of now, the virtual filesystem is incredibly limited (and
in fact might even exhibit some issues with jailbreaking!). I hope to
soon be able to create a file-server likely to be backed by some kind
of embedded database that allows multiprocess access, but also allows me
to store not-quite POSIX file attributes (allowing me to write as a fake
UID or with fake c/a/u times).